created dish, order

+++ db/migrate/20200426083100_create_orders.rb
@@ -1,7 +1,7 @@
 class CreateOrders < ActiveRecord::Migration[5.2]
   def change
     create_table :orders do |t|
-      t.numeric :table_no
+      t.integer :table_no
       t.boolean :is_active
 
       t.timestamps
